+++ PolygonChecker/rectangleSolver.h
@@ -2,6 +2,6 @@
 void sort4PointsXY(int points_x[], int points_y[], int sorted_x[], int sorted_y[]);
 void sort4PointsCCW(int points_x[], int points_y[], int sorted_points[]);
 char* isRectangle(int points_x[], int points_y[], int sorted_points[]);
-float dOfPoints(int pointP[], int pointQ[]);
-float shapePerimeter(int PointA[], int PointB[], int PointC[], int PointD[]);
-float RectArea(int PointA[], int PointB[], int PointC[]);
+double dOfPoints(int pointP[], int pointQ[]);
+double shapePerimeter(int PointA[], int PointB[], int PointC[], int PointD[]);
+double RectArea(int PointA[], int PointB[], int PointC[]);
